+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+### Changed
+- !commands command list is dynamically generated
+- Moderator checks for commands inside get unified
+
 ## [v0.5.4.1-b] - 2019-01-20
 - Updated dependencies, Now vip users are read